LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-audioparam-interface/moderate-exponentialRamp-expected.txt


FAIL Test exponentialRampToValueAtTime() with a moderate ratio of change assert_approx_equals: scheduled value at 22 expected 4.264892619501304 +/- 0.000004067318553448967 but got 4.264896869659424
